Add compile-time feature for desync-tests
