+++ testing/jormungandr-scenario-tests/Cargo.toml
@@ -5,6 +5,10 @@ authors = [ "dev@iohk.io" ]
 edition = "2018"
 default-run = "jormungandr-scenario-tests"
 
+[features]
+# Requires the installation of libfaketime on the host machine
+desync-tests = []
+
 # See more keys and their definitions at https://doc.rust-lang.org/cargo/reference/manifest.html
 
 [dependencies]
