+++ testing/jormungandr-scenario-tests/src/scenario/repository/mod.rs
@@ -176,7 +176,12 @@ fn scenarios_repository() -> Vec<Scenario> {
             transaction_to_passive,
             vec![Tag::Short],
         ),
-        Scenario::new("bft_forks", bft_forks, vec![Tag::Short]),
+        #[cfg(feature = "desync-tests")]
+        Scenario::new(
+            "bft_forks",
+            crate::test::non_functional::desync::bft_forks,
+            vec![Tag::Short],
+        ),
         Scenario::new("interactive", interactive, vec![Tag::Interactive]),
         Scenario::new("example", scenario_2, vec![Tag::Example]),
         Scenario::new("leader_restart", leader_restart, vec![Tag::Short]),
