+++ testing/jormungandr-scenario-tests/src/test/non_functional/mod.rs
@@ -1,3 +1,5 @@
+#[cfg(feature = "desync-tests")]
+pub mod desync;
 pub mod disruption;
 pub mod soak;
